Another Step in Example
* HTML_CONSTANTS: Expanded with more constants (tag names, etc)
* HTML_HELPER: New class that captures some helper functions against the
SHARED_WSF_PERCENT_ENCODER
* ECF: Added libraries: default_nino, encoder, http, wsf

NOTE The ECF needs paths to be corrected from my specific to relative or
better.

* INI: Added the ewf.ini file to ensure one can run the test server.
* HTML_FACTORY: Rearrangement of features with better grouping as well
as some new features that simplify matters like calls for <form>,
<input> and so on.
* HTML_GLOBAL_ATTRIBUTES: changed the nature of how constants are
brought in.
* APPLICATION: Expanded with new code to duplicate the "hello" tutorial
in the eHTML test target as a way to see the eHTML in action.
* HTML_FACTORY_TEST_SET: Changes due to renaming
* TEST_*: Added a number of class-based HTML page generators, which
utilize the HTML_FACTORY, but for specific pages, building them
on-the-fly as-needed by the APPLICATION.