Expanded Table Test
* Expanded the table testing for a larger version beyond the two-month
version from W3-Schools.
* Also added additional require and ensure contracts on the creation
procedure for the TEST_SAVINGS_TABLE class.

NOTE: The plan is for each new test to show/prove the generation
capacity of the eHTML library to quickly design and assemble
HTML_ELEMENT components. The next step beyond that is to bring in CSS-3
and finally Java-script in the full!