Dummies Pages 1 and 2
* Changed name from MOCK to DUMMIES
* Added another DUMMIES page from Ch 4 Basic Lists

NOTE: This is a process for the long haul, where the goal is to mimic
all of the pages shown in this training manual. The hope is that by the
end, the eHTML library will be very well versed in the capacity for
create HTML5, CSS3, and JS pages like a pro.

At some point along the way, I will redo or mimic the Vista Print html
page for Savannah Software Craftsmen and replace it with my own page.
This way, I can use the page as a portfolio piece for Eiffel's ability
to create complex websites with an Eiffel backend and delivery system.